feat(infra): use terraform cloud (#525)

+++ infra/terraform/eks-lepton/uninstall.sh
@@ -6,8 +6,15 @@ if [ -z "$CLUSTER_NAME" ]; then
   CLUSTER_NAME=$(terraform output -json | jq -r '.cluster_name.value')
 fi
 
+if [[ -z $TF_API_TOKEN ]]; then
+  echo "ERROR: Terraform Cloud API token not specified"
+  exit 1
+fi
+
+export "TF_WORKSPACE"=$CLUSTER_NAME
+
 if [ -z "$CLUSTER_NAME" ] || [ "$CLUSTER_NAME" == "null" ]; then
-  terraform destroy -auto-approve
+  echo "ERROR: Cluster name not specified"
 else
   terraform destroy -auto-approve -var="cluster_name=$CLUSTER_NAME"
 fi
